pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "affiliate_outreach_system"
version = "0.1.0"
description = "Affiliate Outreach System for managing and automating outreach campaigns"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "alembic>=1.7.0",
    "celery>=5.1.0",
    "redis>=4.0.0",
    "pydantic>=1.8.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.5",
    "requests>=2.26.0",
    "aiohttp>=3.8.0",
    "prometheus-client>=0.11.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0",
    "boto3>=1.18.0",
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "pytest-asyncio>=0.15.0",
    "black>=21.7b0",
    "isort>=5.9.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "pytest-asyncio>=0.15.0",
    "black>=21.7b0",
    "isort>=5.9.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --cov=app --cov-report=term-missing" 